西元2024年02月28日 (週三) 06時47分30秒 CST

+++ install.sh
@@ -1,11 +1,19 @@
 #!/bin/bash
 
-apt-get update
-apt-get install -y curl wget
+if ! command -v curl &> /dev/null; then
+  apt-get update
+  apt-get install -y curl wget
+fi
 
-wget https://github.com/cloudflare/cloudflared/releases/latest/download/cloudflared-linux-amd64.deb
-dpkg -i cloudflared-linux-amd64.deb
-rm -rf cloudflared-linux-amd64.deb
+# =================================================================
+
+if ! command -v cloudflared &> /dev/null; then
+  wget https://github.com/cloudflare/cloudflared/releases/latest/download/cloudflared-linux-amd64.deb
+  dpkg -i cloudflared-linux-amd64.deb
+  rm -rf cloudflared-linux-amd64.deb
+fi
+
+# =================================================================
 
 rm *.sh
 
@@ -13,6 +21,8 @@ wget https://pulipulichen.github.io/delopy-tunnel/assets/startup.sh
 wget https://pulipulichen.github.io/delopy-tunnel/assets/random_sleep_startup.sh
 
 chmod +x *.sh
+
+# =================================================================
 
 timedatectl set-timezone Asia/Taipei
 
@@ -29,5 +39,7 @@ check_and_append_crontab() {
 
 check_and_append_crontab "@reboot root /root/startup.sh"
 check_and_append_crontab "0 3 * * * root /root/random_sleep_startup.sh"
+
+# =================================================================
 
 reboot
